requirements.txt
django
djangorestframework
confluent-kafka
psycopg2-binary
requests
